chore: move read-local-file utils to lib
