+++ src/lib/read-local-filepaths.ts
