+++ src/views/certifications/content/utils/index.ts
@@ -1,3 +1,1 @@
 export * from './get-certifications'
-export * from './read-local-file'
-export * from './read-local-filepaths'
